requirements.txt
alembic==1.4.2
apache-airflow==1.10.10
apispec==1.3.3
appdirs==1.4.4
appnope==0.1.0
argcomplete==1.11.1
attrs==19.3.0
autopep8==1.5.3
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.1
boto3==1.14.7
botocore==1.17.7
cached-property==1.5.1
cattrs==0.9.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
crayons==0.3.1
croniter==0.3.34
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
dnspython==1.16.0
docutils==0.15.2
email-validator==1.1.1
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.3
flask-swagger==0.2.13
Flask-WTF==0.14.3
fsspec==0.7.4
funcsigs==1.0.2
future==0.18.2
graphviz==0.14
gunicorn==19.10.0
idna==2.9
importlib-metadata==1.6.1
install==1.3.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.10.3
jmespath==0.10.0
json-merge-patch==0.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
lazy-object-proxy==1.5.0
lockfile==0.12.2
lxml==4.5.1
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
natsort==7.0.1
numpy==1.18.5
pandas==0.25.3
parso==0.7.0
pendulum==1.4.4
pexpect==4.8.0
pickleshare==0.7.5
prison==0.1.3
prompt-toolkit==3.0.5
psutil==5.7.0
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyrsistent==0.16.0
python-daemon==2.1.2
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.1.0
pytz==2020.1
pytzdata==2019.3
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.24.0
s3fs==0.4.2
s3transfer==0.3.3
selenium==3.141.0
setproctitle==1.1.10
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.6
tabulate==0.8.7
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.2
thrift==0.13.0
toml==0.10.1
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
typing-extensions==3.7.4.2
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.9
wcwidth==0.2.4
webdriver-manager==3.2.1
Werkzeug==0.16.1
WTForms==2.3.1
zipp==3.1.0
zope.deprecation==4.4.0
